Update Vue tooltip component
