+++ packages/vue/src/tooltip/stories/TooltipDemo.vue
@@ -1,10 +1,7 @@
 <script setup lang="ts">
-import { onMounted, ref } from 'vue'
-
 import Styled from './Styled.vue'
-
-// import Controlled from './Controlled.vue'
-// import CustomDurations from './CustomDurations.vue'
+import Controlled from './Controlled.vue'
+import CustomDurations from './CustomDurations.vue'
 
 export interface OkuTooltipProps {
   template?: 'Styled' | 'Controlled' | 'CustomDurations'
@@ -15,10 +12,6 @@ withDefaults(defineProps<OkuTooltipProps>(), {
   template: 'Styled',
 })
 
-const TooltipRef = ref()
-onMounted(() => {
-  console.warn(TooltipRef.value)
-})
 </script>
 
 <template>
