+++ packages/vue/src/use-composable/useComponentRef.ts
@@ -12,12 +12,16 @@ export function useComponentRef<T = HTMLElement>(forwordEl = false) {
     const el = ['#comment', '#text'].includes(componentRef.value?.$el?.nodeName)
       ? componentRef.value?.$el
       : unrefElement(componentRef)
-    if (el && forwordEl)
+
+    if (el && forwordEl && el instanceof HTMLElement) {
       forword(el)
+      return el
+    }
+    else {
+      return undefined
+    }
     // console.log('componentRef.value?.$el.nodeName', test)
     // console.log(componentRef.value?.$el?.nodeName)
-
-    return el
   })
 
   return {
